corrigindo forma de criar cliente

+++ clientes/views.py
@@ -15,11 +15,7 @@ class ClienteListCreateView(generics.ListCreateAPIView):
     serializer_class = ClienteSerializer
 
     def perform_create(self, serializer):
-        serializer_end = EnderecoSerializer(data=serializer.validated_data['endereco'])
-        serializer_end.is_valid(raise_exception=True)
-        end = serializer_end.save()
-        serializer.is_valid(raise_exception=True)
-        serializer.save(criado_por=self.request.user, ativo=True, endereco=end)
+        serializer.save(criado_por=self.request.user, ativo=True)
         
 
 class ClienteDetailView(generics.RetrieveUpdateDestroyAPIView):
@@ -29,21 +25,12 @@ class ClienteDetailView(generics.RetrieveUpdateDestroyAPIView):
     queryset = Cliente.objects.all()
     serializer_class = ClienteSerializer
 
-    def update(self, request, *args, **kwargs):
-        partial = kwargs.pop('partial', False)
-        instance = self.get_object()
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
-        serializer_end = EnderecoSerializer(instance.endereco, data=request.data, partial=partial)
-        print(serializer.initial_data, serializer_end.initial_data)
-        # serializer.is_valid(raise_exception=True)
-        # serializer_end.is_valid(raise_exception=True)
-        # self.perform_update(serializer_end)
-        # self.perform_update(serializer)
-        # return Response(serializer.data)
-
     def perform_destroy(self, instance):
         try:
-            instance.endereco.delete()
+            if instance.endereco is not None:
+                instance.endereco.delete()
+            if instance.end_entrega is not None:
+                instance.end_entrega.delete()
             instance.delete()
         except ProtectedError as error:
             raise ValidationError(error)
